requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
pydantic
typing-extensions
numpy
requests
python-multipart
streamlit
onnxruntime
Pillow
opencv-python
torch
torchvision
matplotlib
scipy
groq